pyproject.toml
[tool.isort]
profile = "black"
multi_line_output = 3

[tool.poetry]
name = "umailer"
version = "0.1.0"
description = "A FastAPI handler to download and save the registry attachements from emails."
authors = ["Joe Madibekov <jmadibekov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.65.1"
uvicorn = "^0.13.4"
pre-commit = "^2.13.0"
python-dotenv = "^0.17.1"
IMAPClient = "^2.2.0"
chardet = "^4.0.0"
pytz = "^2021.1"
SQLAlchemy = "^1.4.22"
async-exit-stack = "^1.0.1"
async_generator = "^1.10"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
